# laravel/src/vumc-picture-api/storage/api-docs/swagger.yaml
swagger: '2.0'
info:
  description: >-
    Server for the VUMC PICTURE tool to upload, list and share brain maps.
    Created by Jorrit van den Berg from Active Collective
  version: 1.0.0
  title: VUMC PICTURE API
  termsOfService: 'http://tool.pictureproject.nl/terms/'
  contact:
    email: jorrit@activecollective.nl
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'pictureproject.nl:8081'
basePath: /api
...
schemes:
  - http
paths:
  /user:
    get:
      tags:
        - user
      summary: get profile info of the authenticated user
      description: ''
      operationId: getUserProfile
      produces:
        - application/json
      responses:
        '200':
          $ref: '#/definitions/User'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
      security:
        - api_key: []
    post:
      tags:
        - user
      summary: register new user
      description: ''
      operationId: doRegister
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: institute
          in: formData
          description: the institution the user is affiliated to
          required: true
          type: string
          maxLength: 30
        - name: name
          in: formData
          description: full name of the user
          required: true
          type: string
          maxLength: 30
        - name: email
          in: formData
          description: email
          required: true
          type: string
          maxLength: 40
        - name: password
          in: formData
          description: password
          required: true
          type: string
          minLength: 8
          maxLength: 40
        - name: vPassword
          in: formData
          description: verification password
          required: true
          type: string
          minLength: 8
          maxLength: 40
      responses:
        '200':
          description: user registered
          schema:
            $ref: '#/definitions/User'
        '400':
          $ref: '#/responses/BadRequestError'
        '409':
          $ref: '#/responses/ConflictError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
    put:
      tags:
        - user
      summary: update profile info of the authenticated user
      description: ''
      operationId: updateUserProfile
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: institute
          in: formData
          description: the institution the user is affiliated to
          required: false
          type: string
          maxLength: 30
        - name: name
          in: formData
          description: full name of the user
          required: false
          type: string
          maxLength: 30
        - name: email
          in: formData
          description: email
          required: false
          type: string
          maxLength: 40
      responses:
        '200':
          $ref: '#/definitions/User'
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
      security:
        - api_key: []
    delete:
      tags:
        - user
      summary: delete the user profile
      description: ''
      operationId: deleteUserProfile
      produces:
        - application/json
      responses:
        '200':
          description: OK
        '401':
          $ref: '#/responses/UnauthorizedError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
      security:
        - api_key: []
  '/user/activate/{token}':
    get:
      tags:
        - user
      summary: validate the email address with the supplied token
      description: ''
      operationId: registerActivate
      produces:
        - application/json
      parameters:
        - name: token
          in: path
          description: the activation token sended to the user
          required: true
          type: string
      responses:
        '200':
          description: user object
          schema:
            $ref: '#/definitions/User'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
  /user/resend-activation-mail:
    post:
      tags:
        - user
      summary: send a new activation mail
      description: ''
      operationId: resendActivationMail
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: email
          in: formData
          description: the email address of the user
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/responses/BadRequestError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
  /user/login:
    post:
      tags:
        - user
      summary: login a user
      description: ''
      operationId: doLogin
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: email
          in: formData
          description: email
          required: true
          type: string
          maxLength: 40
        - name: password
          in: formData
          description: password
          required: true
          type: string
          minLength: 8
          maxLength: 40
      responses:
        '200':
          description: user logged in
          schema:
            $ref: '#/definitions/User'
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
...
  /user/change-password:
    post:
      tags:
        - user
      summary: change the password of the authenticated user
      description: ''
      operationId: changePassword
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: oldPassword
          in: formData
          description: the existing password
          required: true
          type: string
          maxLength: 40
        - name: newPassword
          in: formData
          description: the new password
          required: true
          type: string
          minLength: 8
          maxLength: 40
        - name: vPassword
          in: formData
          description: verification password
          required: true
          type: string
          minLength: 8
          maxLength: 40
      responses:
        '200':
          description: password changed
          schema:
            $ref: '#/definitions/User'
        '400':
          $ref: '#/responses/BadRequestError'
        '409':
          $ref: '#/responses/ConflictError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
      security:
        - api_key: []
  /user/request-password-reset:
    post:
      tags:
        - user
      summary: send a password reset mail
      description: ''
      operationId: requestPasswordReset
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: email
          in: formData
          description: the email address of the user
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/responses/BadRequestError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
  '/user/find-password-reset/{token}':
    get:
      tags:
        - user
      summary: verify the token from email
      description: ''
      operationId: findPasswordResetRequest
      produces:
        - application/json
      parameters:
        - name: token
          in: path
          description: the token sended to the user
          required: true
          type: string
      responses:
        '200':
          description: password reset object
          schema:
            $ref: '#/definitions/PasswordReset'
        '400':
          $ref: '#/responses/BadRequestError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
  /user/reset-password:
    post:
      tags:
        - user
      summary: perform password reset
      description: ''
      operationId: resetPassword
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: token
          in: formData
          description: the password reset token
          required: true
          type: string
        - name: email
          in: formData
          description: the email address of the user
          required: true
          type: string
        - name: newPassword
          in: formData
          description: the new password
          required: true
          type: string
          minLength: 8
          maxLength: 40
        - name: vPassword
          in: formData
          description: verification password
          required: true
          type: string
          minLength: 8
          maxLength: 40
      responses:
        '200':
          description: user object
          schema:
            $ref: '#/definitions/User'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '429':
          $ref: '#/responses/TooManyRequestsError'
...
responses:
  BadRequestError:
    description: 'bad request, input values missing or invalid'
  UnauthorizedError:
    description: unauthorized request
    headers:
      WWW_Authenticate:
        type: string
...
  ConflictError:
    description: >-
      trying to create an object that already exists or updating a variable with
      the same value
  TooManyRequestsError:
    description: rate limit for endpoint was hit
securityDefinitions:
  api_key:
    type: apiKey
    name: Authorization
    in: header
definitions:
  User:
    type: object
    properties:
      id:
        type: number
        description: the user id
        example: '623594'
      name:
        type: string
        description: full name of the user
        example: John Doe
      institute:
        type: string
        description: the institution the user is working at
        example: VUmc
      email:
        type: string
        description: email address of the user
        example: john.doe@example.com
      apiKey:
        type: string
        description: >-
          the api_key to perform requests with, only added in doLogin,
          doRegister and changePassword operations
        example: gK2xh02ebzc03kafu8ajdpsTT2erdmlef
  PasswordReset:
    type: object
    properties:
      id:
        type: integer
        description: the id of the password reset
        example: 2
      email:
        type: string
        description: the email address of the user
        example: someone@example.com
      token:
        type: string
        description: the password reset token
        example: 32fidmk111m0dwqkmdk1212
      created_at:
        type: string
        description: the datetime on which the password reset object was created
        example: '2018-10-27 12:05:21'
      updated_at:
        type: string
        description: >-
          the datetime on which the password reset request was updated the last
          time
        example: '2018-10-27 12:05:21'
...